dedupe `waitForHashChangeEvent`
